# Generate API validation instead of using reflection