--- a/models/protocol_tests/generate.go
+++ b/models/protocol_tests/generate.go
@@ -54,6 +54,7 @@ var _ = util.Trim("")
 var _ = url.Values{}
 var _ = io.EOF
 var _ = aws.String
+var _ = fmt.Println
 
 func init() {
 	protocol.RandReader = &awstesting.ZeroReader{}
@@ -72,6 +73,7 @@ var extraImports = []string{
 	"bytes",
 	"encoding/json",
 	"encoding/xml",
+	"fmt",
 	"io",
 	"io/ioutil",
 	"net/http",