--- a/private/model/api/api_test.go
+++ b/private/model/api/api_test.go
@@ -1,3 +1,5 @@
+// +build 1.6
+
 package api
 
 import (