--- a/private/model/api/load.go
+++ b/private/model/api/load.go
@@ -60,5 +60,9 @@ func (a *API) Setup() {
 		a.removeUnusedShapes()
 	}
 
+	if !a.NoValidataShapeMethods {
+		a.addShapeValidations()
+	}
+
 	a.initialized = true
 }