--- a/private/model/api/load_test.go
+++ b/private/model/api/load_test.go
@@ -1,3 +1,5 @@
+// +build 1.6
+
 package api
 
 import (