… aws/corehandlers/param_validator_test.go
… private/model/api/shape.go
… private/model/api/shapetag_test.go
… private/protocol/ec2query/build_test.go
… private/protocol/ec2query/unmarshal_test.go
… private/protocol/jsonrpc/build_test.go
… private/protocol/jsonrpc/unmarshal_test.go
… private/protocol/query/build_test.go
… private/protocol/query/unmarshal_test.go
… private/protocol/restjson/build_test.go
… private/protocol/restjson/unmarshal_test.go
… private/protocol/restxml/build_test.go
… private/protocol/restxml/unmarshal_test.go
… service/acm/api.go
… service/apigateway/api.go
… service/autoscaling/api.go
… service/cloudformation/api.go
… service/cloudfront/api.go
… service/cloudhsm/api.go
… service/cloudsearch/api.go
… service/cloudsearchdomain/api.go
… service/cloudtrail/api.go
… service/cloudwatch/api.go
… service/cloudwatchevents/api.go
… service/cloudwatchlogs/api.go
… service/codecommit/api.go
… service/codedeploy/api.go
… service/codepipeline/api.go
… service/cognitoidentity/api.go
… service/cognitoidentityprovider/api.go