Updates the SDK to generate its input parameter validation logic instead
of using reflection. This improves the performance (cpu/memory) of
building requests by about 15-20%.

This change should be transparent to users of the SDK, other than reduction of memory usage, and delay caused by param validation. This change does also refactor the SDK's internal API generator's templates to require go1.6. This will not impact usage of the SDK, just its code generation, which is covered by the [private pkg pact](https://github.com/aws/aws-sdk-go/blob/master/private/README.md).
   
Fix #537

Suggested to not review the diff as a whole at once since the generated services/protocol commit is pretty large.

This PR greatly improves the performance of validating API operation params. At the micro level the impact of validation now is extremely little. In the context of marshaling and signing a request this improves validation by about 15%-20%. 

`go test -bench . -benchmem -run NONE ./aws/corehandlers/`
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkValidateAny-4     203116        790           -99.61%

benchmark                  old allocs     new allocs     delta
BenchmarkValidateAny-4     1909           0              -100.00%

benchmark                  old bytes     new bytes     delta
BenchmarkValidateAny-4     19363         0             -100.00%
```

Benchmark results with a simpler operation, and additional isolated bench of validation by its self.

[`go test -bench . -run NONE -benchmem bench_test.go`](https://gist.github.com/jasdel/ca7e97be2ee3e3e7f616efa04d445f64)
```
benchmark                old ns/op     new ns/op     delta
BenchmarkOperation-4     77574         66292         -14.54%
BenchmarkValidate-4      12554         36.4          -99.71%

benchmark                old allocs     new allocs     delta
BenchmarkOperation-4     494            421            -14.78%
BenchmarkValidate-4      73             0              -100.00%

benchmark                old bytes     new bytes     delta
BenchmarkOperation-4     24571         24135         -1.77%
BenchmarkValidate-4      440           0             -100.00%